+++ aws_account/02_core/main.tf
@@ -63,62 +63,63 @@ module "security_global" {
 #   tags    = var.tags
 # }
 
-# # Create the VPC with:
-# # - public and private subnets
-# # - NAT gateway, Internet gateway, Network ACLs
-# # - empty default security group
-# # - "main" security group as a starting point (HTTP, HTTPS, SSH)
-# # - VPC flow logs in CloudWatch
-# module "myvpc" {
-#   source             = "../../modules/vpc"
-#   name               = "main"
-#   region             = data.aws_region.current.name
-#   kms_key            = data.aws_kms_key.main
-#   cidr               = "10.2.0.0/16"
-#   log_bucket         = data.aws_s3_bucket.logging
-#   log_retention_days = var.log_retention_days
-#   vpc_endpoints      = false
-#   subnets_public = [
-#     {
-#       name = "main"
-#       az   = "a"
-#       cidr = "10.2.1.0/24"
-#     },
-#     {
-#       name = "main"
-#       az   = "b"
-#       cidr = "10.2.2.0/24"
-#     },
-#     {
-#       name = "main"
-#       az   = "c"
-#       cidr = "10.2.3.0/24"
-#     }
-#   ]
-#   subnets_private = [
-#     {
-#       name = "main"
-#       az   = "a"
-#       cidr = "10.2.11.0/24"
-#     },
-#     {
-#       name = "main"
-#       az   = "b"
-#       cidr = "10.2.12.0/24"
-#     },
-#     {
-#       name = "main"
-#       az   = "c"
-#       cidr = "10.2.13.0/24"
-#     },
-#     # {
-#     #   name = "vpn"
-#     #   az   = "a"
-#     #   cidr = "10.2.99.0/24"
-#     # }
-#   ]
-#   tags = var.tags
-# }
+# Create the VPC with:
+# - public and private subnets
+# - NAT gateway, Internet gateway, Network ACLs
+# - empty default security group
+# - "main" security group as a starting point (HTTP, HTTPS, SSH)
+# - VPC flow logs in CloudWatch
+module "myvpc" {
+  source             = "../../modules/vpc"
+  name               = "main"
+  region             = data.aws_region.current.region
+  kms_key            = data.aws_kms_key.main
+  cidr               = "10.2.0.0/16"
+  log_bucket         = data.aws_s3_bucket.logging
+  log_retention_days = var.log_retention_days
+  subnets_public = [
+    {
+      name = "main"
+      az   = "a"
+      cidr = "10.2.1.0/24"
+    },
+    {
+      name = "main"
+      az   = "b"
+      cidr = "10.2.2.0/24"
+    },
+    {
+      name = "main"
+      az   = "c"
+      cidr = "10.2.3.0/24"
+    }
+  ]
+  subnets_private = [
+    {
+      name = "main"
+      az   = "a"
+      cidr = "10.2.11.0/24"
+    },
+    {
+      name = "main"
+      az   = "b"
+      cidr = "10.2.12.0/24"
+    },
+    {
+      name = "main"
+      az   = "c"
+      cidr = "10.2.13.0/24"
+    },
+    # {
+    #   name = "vpn"
+    #   az   = "a"
+    #   cidr = "10.2.99.0/24"
+    # }
+  ]
+  tags = var.tags
+  # Warning: Disabling Dev Mode will be costly
+  dev_mode = true
+}
 
 # # External Load Balancer (Public)
 # module "alb_public" {