+++ modules/vpc/vpc_endpoints_cost.tf
@@ -1,22 +1,11 @@
 
-# Documentation
+# Documentation for SSM VPC Endpoints
 # https://docs.aws.amazon.com/systems-manager/latest/userguide/setup-create-vpc.html
 
-# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/vpc_endpoint
-resource "aws_vpc_endpoint" "gateways" {
-  for_each          = toset(var.vpc_endpoints == true ? ["s3", "dynamodb"] : [])
-  vpc_id            = aws_vpc.main.id
-  service_name      = "com.amazonaws.${data.aws_region.current.region}.${each.value}"
-  vpc_endpoint_type = "Gateway"
-  route_table_ids   = [aws_default_route_table.private.id, aws_route_table.public.id]
-  tags = merge(var.tags, {
-    Name = "vpc-endpoint-${each.value}"
-  })
-}
-
+# These are about $8/mo/endpoint/az (e.g., $8/mo * 10 Services * 3 AZs = ~$240/mo)
 # https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/vpc_endpoint
 resource "aws_vpc_endpoint" "interfaces" {
-  for_each            = toset(var.vpc_endpoints == true ? ["kms", "logs", "secretsmanager", "ssm", "ec2messages", "ec2", "ssmmessages"] : [])
+  for_each            = toset(var.dev_mode ? [] : var.vpc_endpoints_interface)
   vpc_id              = aws_vpc.main.id
   service_name        = "com.amazonaws.${data.aws_region.current.region}.${each.value}"
   vpc_endpoint_type   = "Interface"
@@ -26,13 +15,14 @@ resource "aws_vpc_endpoint" "interfaces" {
     aws_security_group.vpce[0].id,
   ]
   tags = merge(var.tags, {
-    Name = "vpc-endpoint-${each.value}"
+    Name = "vpce-${each.value}"
   })
 }
 
+# Security Groups only apply to Interface VPC Endpoints
 # https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/security_group
 resource "aws_security_group" "vpce" {
-  count       = var.vpc_endpoints ? 1 : 0
+  count       = var.dev_mode ? 0 : 1
   name        = "vpc-endpoints"
   description = "VPC Endpoints"
   vpc_id      = aws_vpc.main.id
@@ -41,7 +31,7 @@ resource "aws_security_group" "vpce" {
     Name = "vpc-endpoints"
   })
 
-  # HTTP from Anywhere
+  # HTTPS from VPC
   # Use only for redirecting HTTPS (443)
   ingress {
     from_port   = 443