+++ modules/vpc/vpc_endpoints_free.tf
@@ -1,0 +1,12 @@
+
+# https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/vpc_endpoint
+resource "aws_vpc_endpoint" "gateways" {
+  for_each          = toset(["s3", "dynamodb"])
+  vpc_id            = aws_vpc.main.id
+  service_name      = "com.amazonaws.${data.aws_region.current.region}.${each.value}"
+  vpc_endpoint_type = "Gateway"
+  route_table_ids   = [aws_default_route_table.private.id, aws_route_table.public.id]
+  tags = merge(var.tags, {
+    Name = "vpce-${var.name}-${each.value}"
+  })
+}